requirements.txt
datasets
librosa
soundfile
gradio-client
git+https://github.com/unitaryai/detoxify
pyloudnorm
langdetect
pydub
